+++ packages/orq-rc/src/orq_ai_sdk/models/listcontactsop.py
@@ -184,7 +184,7 @@ class Data(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-10-14T05:25:41.367Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-14T11:49:00.661Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")