+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K7GJX5YDHYDVWB3JT4R8VKE5"
+        "01K7H8V1KXZWC02DF6S89F9FTX"
     )
     r"""The id of the resource"""